Rename app.py to web.py
